backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
joblib
python-dotenv